Feng compatibility

+++ shaders/skinned.vert
@@ -4,7 +4,7 @@ precision lowp float;
 layout(location=0) in vec3 pos;
 layout(location=1) in vec3 normal;
 layout(location=2) in vec2 tx;
-layout(location=3) in ivec4 vertIndices;
+layout(location=3) in vec4 vertIndices;
 layout(location=4) in vec4 vertWeights;
 
 //separate model matrix for normals
@@ -18,19 +18,19 @@ out vec3 fragPos; //world space position
 out vec3 modNormal;
 out vec2 texCoord;
 
-vec3 deformVertLBS(vec3 target, ivec4 indices,vec4 weights){
+vec3 deformVertLBS(vec3 target, vec4 indices,vec4 weights){
     vec3 result=vec3(.0);
     for(int i=0;i<4;i++){
-        result+=weights[i]*poses[indices[i]]*vec4(target,1.0);
+        result+=weights[i]*poses[int(indices[i])]*vec4(target,1.0);
     }
     return result;
 }
 
-vec3 deformNormalLBS(vec3 target, ivec4 indices,vec4 weights){
+vec3 deformNormalLBS(vec3 target, vec4 indices,vec4 weights){
     vec3 result=vec3(.0);
     for(int i=0;i<4;i++){
 
-        result+=weights[i]*mat3(poses[indices[i]])*target;
+        result+=weights[i]*mat3(poses[int(indices[i])])*target;
     }
     return result;
 }
